new way to declare C routines
